init react routing frontend architecture
