+++ client/package.json
@@ -22,27 +22,28 @@
   },
   "homepage": "http://ludobox.net",
   "devDependencies": {
-      "babel": "5.8.29",
-      "babel-cli": "^6.18.0",
-      "babel-core": "^6.18.2",
-      "babel-loader": "^6.2.7",
-      "babel-preset-es2015": "^6.18.0",
-      "babel-preset-react": "^6.16.0",
-      "babel-preset-stage-0": "^6.16.0",
-      "css-loader": "^0.25.0",
-      "gh-pages": "^0.11.0",
-      "json-loader": "^0.5.4",
-      "nodemon": "1.8.0",
-      "react-hot-loader": "^3.0.0-beta.6",
-      "redux-logger": "^2.7.4",
-      "style-loader": "^0.13.1",
-      "webpack": "^1.13.3",
-      "webpack-dev-server": "^1.16.2"
+    "babel": "5.8.29",
+    "babel-cli": "^6.18.0",
+    "babel-core": "^6.18.2",
+    "babel-loader": "^6.2.7",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
+    "babel-preset-stage-0": "^6.16.0",
+    "css-loader": "^0.25.0",
+    "gh-pages": "^0.11.0",
+    "json-loader": "^0.5.4",
+    "nodemon": "1.8.0",
+    "react-hot-loader": "^3.0.0-beta.6",
+    "redux-logger": "^2.7.4",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.13.3",
+    "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
     "axios": "^0.15.3",
     "react": "^15.3.2",
-    "react-dom": "^15.3.2"
+    "react-dom": "^15.3.2",
+    "react-router": "^3.0.2"
   },
   "scripts": {
     "start": "NODE_ENV=development webpack --watch",
