+++ client/src/components/404/404.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+export default class Page404 extends React.Component {
+
+  render() {
+    return (
+      <p>
+        <h1>Ludobox</h1>
+        <p>Wooops... nothing here !</p>
+      </p>
+    )
+  }
+}
