+++ client/src/components/About/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default class AboutPage extends React.Component {
+
+  render() {
+    console.log("olala!");
+    return (
+      <h1>About</h1>
+    )
+  }
+}
