+++ client/src/components/AddGame/AddGame.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+export default class AddGame extends React.Component {
+
+  render() {
+    return (
+      <span>
+        <h1>Ludobox</h1>
+        <p>Add new game !</p>
+      </span>
+    )
+  }
+}
