+++ client/src/components/Home/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default class Home extends React.Component {
+
+  render() {
+    return (
+      <span>
+        <p>Welcome to Ludobox :)</p>
+      </span>
+    )
+  }
+}
